Imeplementation and tests for front- and back-end for TimeLine view (for providing issues).

+++ event-planner/src/test/java/ut/helpers/ActiveObjectsTestHelper.java
@@ -56,7 +56,7 @@ public class ActiveObjectsTestHelper {
         return result;
     }
 
-    public Task createTask(String name, int time) {
+    public Task createTask(String name, long time) {
         Task result = activeObjects.create(Task.class);
         result.setName(name);
         result.setTimeToComplete(time);
